recfactor: chat refactor
- Entity 디렉토리 -> Document 디렉토리로 분리
- 안 쓰는 Import 코드 제거
- ChatMessage Lombok 적용
