+++ src/main/java/com/sparta/levelup_backend/domain/chat/document/ChatMessage.java
@@ -1,13 +1,19 @@
-package com.sparta.levelup_backend.domain.chat.entity;
+package com.sparta.levelup_backend.domain.chat.document;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
 
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 
+@Builder
 @Getter
+@AllArgsConstructor
 @Document(collection = "chat_messages")
 public class ChatMessage {
 
@@ -18,10 +24,6 @@ public class ChatMessage {
 	private String message;
 	private LocalDateTime timestamp;
 
-	public ChatMessage(Long chatroomId, String nickname, String message, LocalDateTime timestamp) {
-		this.chatroomId = chatroomId;
-		this.nickname = nickname;
-		this.message = message;
-		this.timestamp = timestamp;
-	}
+	@Builder.Default
+	private List<String> readUsers = new ArrayList<>();
 }
